# bindings/go: Reduce memory allocations when packing tuples