… bindings/go/src/fdb/tuple/testdata/tuples.golden