@vmg 2018-05-01
Fixed up @alecgrieser!

- Rebased on top of `release-5.2`
- Capitalized `putBytes` et al, renamed `ii` to something more obvious
- Added a test case for nested tuple packing

Thanks for the review. Let me know what else would you like me to tweak!